moved plugin_registry pytest functionality to Unit Test Module
